requirements.txt
# local package
-e .

# Google Cloud Storage client library for GCS operations
google-cloud-storage>=2.0.0

# GCSFS allows Python code to interact with Google Cloud Storage via a filesystem interface
gcsfs>=2021.4.0

# PyYAML for parsing YAML configuration files
PyYAML>=5.4

# Pandas for data manipulation and analysis
pandas>=1.2.0

# scikit-learn for feature engineering, model building, evaluation, and metrics
scikit-learn>=0.24.0

# NLTK for text preprocessing and tokenization
nltk>=3.6.0

# FastAPI for building the serving API for predictions
fastapi>=0.68.0

# Uvicorn as an ASGI server to run FastAPI apps
uvicorn>=0.15.0

# Joblib for model and vectorizer serialization
joblib>=1.0.0

# MLflow for experiment tracking and model registry
mlflow>=1.20.0

# DVC with GCS support for data versioning
dvc[gcs]>=2.0.0

# Pytest for running tests (optional but recommended)
pytest>=6.0.0


python-dotenv